backend/requirements.txt
fastapi
openai
llama-index
pinecone-client
langchain
arxiv
uvicorn